==> 1-1. oliveyoung_product_list_crawling.ipynb <==
# 1-1. 올리브영 선케어 제품 URL 리스트업 크롤링

from selenium import webdriver
from selenium.webdriver.common.by import By

import time
import random
import csv
import re
import pandas as pd

# 선스프레이 URL
url = "https://www.oliveyoung.co.kr/store/display/getMCategoryList.do?dispCatNo=1000001001100050002&fltDispCatNo=&prdSort=01&pageIdx={}&rowsPerPage=24&searchTypeSort=btn_thumb&plusButtonFlag=N&isLoginCnt=0&aShowCnt=0&bShowCnt=0&cShowCnt=0&trackingCd=Cat1000001001100050002_Small&amplitudePageGubun=SMALL_CATE&t_page=%EC%B9%B4%ED%85%8C%EA%B3%A0%EB%A6%AC%EA%B4%80&t_click=%EC%B9%B4%ED%85%8C%EA%B3%A0%EB%A6%AC%EC%83%81%EC%84%B8_%EC%86%8C%EC%B9%B4%ED%85%8C%EA%B3%A0%EB%A6%AC&midCategory=%EC%84%A0%ED%8C%A8%EC%B9%98&smallCategory=%EC%86%8C_%EC%84%A0%EC%8A%A4%ED%94%84%EB%A0%88%EC%9D%B4&checkBrnds=&lastChkBrnd=&t_3rd_category_type=%EC%86%8C_%EC%84%A0%EC%8A%A4%ED%94%84%EB%A0%88%EC%9D%B4"

# 선패치 URL
url = "https://www.oliveyoung.co.kr/store/display/getMCategoryList.do?dispCatNo=1000001001100050001&fltDispCatNo=&prdSort=01&pageIdx={}&rowsPerPage=24&searchTypeSort=btn_thumb&plusButtonFlag=N&isLoginCnt=0&aShowCnt=0&bShowCnt=0&cShowCnt=0&trackingCd=Cat1000001001100050001_Small&amplitudePageGubun=SMALL_CATE&t_page=%EC%B9%B4%ED%85%8C%EA%B3%A0%EB%A6%AC%EA%B4%80&t_click=%EC%B9%B4%ED%85%8C%EA%B3%A0%EB%A6%AC%EC%83%81%EC%84%B8_%EC%86%8C%EC%B9%B4%ED%85%8C%EA%B3%A0%EB%A6%AC&midCategory=%EC%84%A0%EC%8A%A4%ED%94%84%EB%A0%88%EC%9D%B4%2F%EC%84%A0%ED%8C%A8%EC%B9%98&smallCategory=%EC%86%8C_%EC%84%A0%ED%8C%A8%EC%B9%98&checkBrnds=&lastChkBrnd=&t_3rd_category_type=%EC%86%8C_%EC%84%A0%ED%8C%A8%EC%B9%98"

# 페이지 클릭 수

driver = webdriver.Chrome()
with open('./sunpatch_link_test.csv', "w", newline="", encoding = "utf-8") as fw:
    writer = csv.DictWriter(fw, fieldnames=['page', 'brand_name', 'product_name', 'product_link', 'image_url'])
    writer.writeheader()
    for i in range(1, 3):
        new_url = url.format(i)
        driver.get(new_url)
        timesleep = random.randint(1, 5)  # 1에서 5초 사이 랜덤 슬립
        time.sleep(timesleep)
        goods = driver.find_elements(By.CLASS_NAME, 'prd_info ')
        for good in goods:
            try:
                # h2 요소 아래의 a 태그 찾기
                product_link = good.find_element(By.TAG_NAME, 'a')
                brand_name = good.find_element(By.CLASS_NAME, 'tx_brand').text
                product_name = good.find_element(By.CLASS_NAME, 'tx_name').text
                img = good.find_element(By.TAG_NAME, 'img')
                imgsrc = img.get_attribute('src')
                writer.writerow({
                    'page': i,
                    'brand_name': brand_name,
                    'product_name': product_name,
                    'product_link': product_link.get_attribute('href'),
                    'image_url': imgsrc
                })
                print(f"brand: {brand_name}, Link: {product_link}")
            except:
                # a 태그가 없는 경우 예외 처리
                print("a 태그를 찾을 수 없음")
                continue   
driver.quit()

df
