==> 2. data_preprocessing_and_tokenization.ipynb <==
# 2. 리뷰 크롤링 데이터 전처리 및 토큰화 작업

from konlpy.tag import Komoran
import pandas as pd
import re

각 카테고리별로 아래 전처리 작업을 수행했습니다.  
- 브랜드명 칼럼 추가
- 카테고리 칼럼 추가
- 평점 칼럼 점수만 추출
- 피부타입 칼럼 리스트에서 대표 피부타입만 추출
- 리뷰 텍스트 정제 및 토큰화

## 1. Suncream

# 선크림 리뷰 데이터 불러오기
suncream_review_df = pd.read_csv("./crawling_data/suncream_reviews.csv",encoding="utf-8")

# 선크림 링크 데이터 불러오기
suncream_link_df = pd.read_csv("./crawling_data/suncream_link.csv",encoding="utf-8")

# 브랜드명 리뷰데이터에 추가해주기
suncream_review_df = pd.merge(suncream_review_df, suncream_link_df[['product_name','brand_name']], on = 'product_name',how='left')

# 평점칼럼 전처리
suncream_review_df['star'] = suncream_review_df['star'].apply(lambda a: int(a.split(" ")[1][0]))

# 피부타입 칼럼 전처리
suncream_review_df['skin_type'] = suncream_review_df['skin_type'].apply(lambda x: eval(x)[0] if isinstance(eval(x), list) and len(eval(x)) > 0 else "")

# 카테고리에 선크림 추가
suncream_review_df['category'] = 'suncream'

# 토큰화를 위해 결측치 채우기
suncream_review_df.fillna("")

# 명사로 구분된는 단어중 인사이트에 의미없는 단어 리스트
suncream_outword = ['전', '사', '합', '당', '갈', '왜', '한', '겨', '편', '을', '김', '날', '안', '후', '넘', '칠', '면', '요', '위', '막',
           '이', '양', '낮', '라', '못', '선', '아', '여', '어', '낙', '영', '첨', '일', '용', '함', '느', '걸', '극', '차', '부', '눈',
           '티', '흠', '뒤', '고', '규', '열', '력', '쌈', '보', '감', '질', '예', '글', '개', '뭐', '다', '애', '달', '황', '심',
           '만', '습', '오', '뜰','갑', '피', '곳', '미', '와', '출', '샷', '시', '폭', '껀', '메', '암', '게', '니', '닷', '쌍', '입',
           '큐', '링', '디', '버', '잘', '하', '유', '정', '난', '박', '길', '불', '드', '남', '띠', '랩', '답', '액', '서', '엄', '청',
           '주', '표', '가', '끔', '배', '끝', '떼', '말', '패', '성', '종', '순', '분', '들', '법', '학', '악', '금', '진', '폴', '솔',
           '자', '잔', '폰', '봉', '음', '밑', '상', '얼', '기', '색', '향', '존', '겉', '구', '꾀', '융', '의', '더', '삶', '광', '천',
           '귀', '우', '파', '르', '직', '에', '최', '공', '씨', '땅', '품', '컷', '지', '득', '은', '리', '괘', '룸', '과', '추', '겁',
           '돌', '실', '알', '강', '녀', '도', '까', '덕', '틈', '두', '담', '조', '으', '봄', '랑', '접', '깨', '핑', '통', '엘', '인',
           '류', '끈', '란', '비', '효', '흉', '혼', '풀', '형', '찬', '단', '홀', '엉', '꼴', '따', '임', '본', '매', '볕', '밭', '포',
           '냉', '평', '테', '팁', '필', '계', '린', '쥐', '로', '총', '붐', '런', '쿠', '착', '림', '동', '뇨', '세', '촉', '반', '모',
           '적', '각', '빰', '야', '급', '노', '놀', '탓', '업', '쥬', '운', '염', '제', '빠', '혀', '화', '낫', '밴', '점', '토', '러',
           '석', '신', '톨', '완', '룰', '나', '절', '탈', '칼', '족', '소', '작', '중', '휴', '재', '래', '졸', '삼', '능', '욥', '살',
           '힐', '방', '문', '처', '벗', '새', '내', '짐', '백', '뉴', '좀', '침', '첩', '마', '관', '끼', '책', '국', '텐', '씀', '널',
           '맨', '톰', '궁', '션', '흑', '몰', '님', '레', '굴', '울', '줄', '저','때', '선크림','썬크림','제품', '구매','사용','느낌','생각']

from tqdm.notebook import tqdm
import re
from konlpy.tag import Komoran


tqdm.pandas()

komoran = Komoran(userdic = "./data/user.dic")
def clean_text(text):# 이모티콘제거
    if isinstance(text, str):
        text = re.sub(r'[^\x00-\x7F\uAC00-\uD7AF]', '', text)
        text = re.sub(r'[!]', '', text)
        return text
    else:
        return ''  

def process_review(sentence):# 토큰화 진행(명사가져오기, outword_list의 단어 빼기)
    try:
        clean_sentence = clean_text(sentence) 
        tokens = komoran.pos(clean_sentence)
        nouns = [x[0] for x in tokens if x[1] in ("NNG", "NNP") and x[0] not in suncream_outword]
        return nouns
    except UnicodeDecodeError as e:
        print(f"UnicodeDecodeError processing sentence: {sentence}")
        print(e)
        return []
    except Exception as e:
        print(f"Error processing sentence: {sentence}")
        print(e)
        return []


suncream_review_df['processed_review'] = suncream_df['review'].progress_apply(process_review)


#파일저장
suncream_review_df.to_csv("./preprocessed_data/preprocessed_suncream_data.csv",encoding="utf-8-sig",index=False)

## 2. Suncushion

# 선쿠션 리뷰 데이터 불러오기
suncushion_review_df = pd.read_csv("./crawling_data/suncushion_reviews.csv",encoding="utf-8")

# 선쿠션 링크 데이터 불러오기
suncushion_link_df = pd.read_csv("./crawling_data/suncushion_link.csv",encoding="utf-8")

# 선쿠션 리뷰 데이터에 브랜드 명 추가
suncushion_review_df = pd.merge(suncushion_review_df, suncushion_link_df[['product_name','brand_name']], on = 'product_name',how='left')

# 평점 전처리
suncushion_review_df['star'] = suncushion_review_df['star'].apply(lambda a: int(a.split(" ")[1][0]))

# 피부타입 칼럼 전처리
suncushion_review_df['skin_type'] = suncushion_review_df['skin_type'].apply(lambda x: eval(x)[0] if isinstance(eval(x), list) and len(eval(x)) > 0 else "")

# 리뷰데이터에 카테고리 열 추가
suncushion_review_df['category'] = 'suncushion'

# 결측치 채우기
suncushion_review_df.fillna("")

# 명사로 구분된는 단어중 인사이트에 의미없는 단어 리스트
suncushion_outword = ["때", "선쿠션", '쌀', '요', '양', '형', '혼', '넘', '톤', '몰', '하', '연', '안', '폰', '티', '목', '립', '킴',
         '종', '전', '본', '가', '미', '다', '게', '아', '혀', '눈', '향', '함', '침', '맘', '후', '맥', '힘', '달', '겉', '선', '열', '땀',
         '날', '말', '밖', '여', '갈', '을', '못', '손', '물', '사', '폼', '입', '이', '당', '족', '파', '뒤', '볼', '코', '끔', '한', '톰',
         '업', '용', '콩', '떡', '허', '팁', '기', '뼈', '임', '끝', '감', '빛', '덧', '싱', '노', '칠', '품', '얼', '오', '위', '일', '틈',
         '리', '습', '속', '도', '뜸', '색', '어', '짜', '피', '발', '영', '남', '쌍', '야', '엉', '앞', '욥', '개', '귀', '썬', '골', '봄',
         '옆', '보', '팔', '인', '꿀', '그', '학', '괘', '은', '병', '해', '졸', '똥', '씀', '정', '두', '딸', '핑', '특', '알', '조', '흠',
         '팥', '합', '막', '잘', '난', '콘', '셀', '통', '누', '무', '더', '과', '팩', '화', '짐', '재', '지', '옷', '린', '모', '광', '글',
         '독', '애', '디', '션', '내', '매', '엽', '떼', '쌈', '뽕', '집', '능', '몸', '소', '추', '친', '완', '첨', '중', '곳', '쥬', '각',
         '느', '잇', '베', '챙', '겨', '서', '덤', '생', '마', '예', '세', '으', '염', '유', '나', '적', '러', '약', '새', '항', '상', '흰',
         '끼', '메', '컵', '악', '흑', '데', '음', '옴', '담', '획', '신', '로', '갑', '면', '단', '읍', '고', '곱', '띠', '엿', '넌', '냉',
         '백', '결', '존', '극', '총', '납', '라', '레', '래', '멜', '호', '턱', '타', '쿠', '뭐', '맛', '찬', '따', '녀', '살', '뺨', '뜰',
         '체', '절', '득', '드', '김', '규', '별', '변', '효', '운', '뭍', '길', '올', '죽', '순', '짝', '버', '릭', '겹', '끈', '냐', "요조",
         '편', '만', '밥', '식', '믈', '주', '풀', '갸', '부', '공', '혹', '력', '춤', '표', '배', '시', '밑', '낮', '진', '벗', '덕', '땅',
         '건', '릴', '돌', '착', '컷', '장', '비', '현', '불', '밤', '반', '니', '닷', '직', '처', '참', '벤', '박', '방', '국', '환', '쥐',
         '르', '룩', '암', '탓', '님', '홈', '패', '답', '차', '급', '점', '꿈', '에', '투', '숨', '강', '꽃', '랑', '까', '징', '걸', '센',
         '언', '왕', '엄', '찌', '링', '빠', '껌', '토', '카', '치', '질', '켄', '왜', '뜻', '구', '성', '좌', '우', '충', '거', '태', '증',
         '의', '저', '톱', '탁', '앎', '테', '섶', '림', '꼴', '팀', '액', '둑', '런', '탑', '룸', '싹', '락', '흉', '샘', '응', '송', '망',
         '홀', '줄', '낫', '필', '빙', '팝', '빅', '작', '문', '민', '실', '심', '텐', '죤', '울', '키', '값', '릉', '철', '동', '금', '팬',
         '홍', '창', '벡', '검', '깐', '칸', '휘', '율', '깃', '랩', '잎', '평', '숲', '논', '콜', '틱', '계', '류', '최', '축', '닐', '탄',
         '힐', '윤', '산', '므', '잠', '헌', '짓', '곰', '택', '낭', '덫', '몽', '역', '란', '들', '샹', '뇨', '핫', '범', '댐', '칩', '록',
         '외', '술', '빵', '궁', '뻘', '폴', '묘', '셤', '녜', '줌', '농', '샷', '삼', '찜', '와', '수', '먹', '펙', '초', '원', '폭', '혁',
         '믹', '법', '끌', '핵', '온', '닻', '청', '멱', '펄', '뤼', '량', '복', '잡', '묵', '칙', '솔', '맨', '잔', '컴', '자', '휴', '밭',
         '좀', '삽', '근', '접', '포', '썰', '땡', '탕', '붓', '펜', '탈', '벨', '솜', '댁', '램', '롤', '봉', '몬', '벽', '틀', '햄', '벼',
         '꾀', '멋', '씨', '껀', '젤', '즙', '활', '석', '꽈', '몫', '뱃', '옥', '녹', '촉', '킥', '엔', '볕', '멀', '쇼', '첩', '터', '겁',
         '젓', '굴', '행', '붕', '훔', '루', '놀', '론', '깨', '솥', '경', '출', '갭', '욕', '갱', '밴', '널', '털', '돗', '쾌', '죄', '붐',
          '구매', '제품', '사용', '느낌', '생각', '시나', '모로', "개도", "데바르", "이면", "이건", "아악", "전이", "호가", "안보",
          "군가", "지나", "리한", "여물", "바쿠", "바르", "쓰기", "하주", "이랑"]

from tqdm.notebook import tqdm
import re
from konlpy.tag import Komoran


tqdm.pandas()

komoran = Komoran(userdic = "./data/user.dic")
def clean_text(text):# 이모티콘제거
    if isinstance(text, str):
        text = re.sub(r'[^\x00-\x7F\uAC00-\uD7AF]', '', text)
        text = re.sub(r'[!]', '', text)
        return text
    else:
        return ''  

def process_review(sentence):# 토큰화 진행(명사가져오기, outword_list의 단어 빼기)
    try:
        clean_sentence = clean_text(sentence)  
        tokens = komoran.pos(clean_sentence)
        nouns = [x[0] for x in tokens if x[1] in ("NNG", "NNP") and x[0] not in suncushion_outword]
        return nouns
    except UnicodeDecodeError as e:
        print(f"UnicodeDecodeError processing sentence: {sentence}")
        print(e)
        return []
    except Exception as e:
        print(f"Error processing sentence: {sentence}")
        print(e)
        return []


suncushion_review_df['processed_review'] = suncushion_review_df['review'].progress_apply(process_review)


#파일저장
suncushion_review_df.to_csv("./preprocessed_data/preprocessed_suncushion_data.csv",encoding="utf-8-sig",index=False)

## 3. Sunstick

# 선스틱 리뷰 데이터 불러오기
sunstick_review_df = pd.read_csv("./crawling_data/sunstic_reviews.csv",encoding="utf-8")

# 선스틱 링크 데이터 불러오기
sunstick_review_df = pd.read_csv("./crawling_data/sunstic_link.csv",encoding="utf-8")

# 브랜드명 리뷰데이터에 추가해주기
sunstick_review_df = pd.merge(sunstick_review_df, sunstick_link_df[['product_name','brand_name']], on = 'product_name',how='left')

# 평점칼럼 전처리
sunstick_review_df['star'] = sunstick_review_df['star'].apply(lambda a: int(a.split(" ")[1][0]))

# 피부타입 칼럼 전처리
sunstick_review_df['skin_type'] = sunstick_review_df['skin_type'].apply(lambda x: eval(x)[0] if isinstance(eval(x), list) and len(eval(x)) > 0 else "")

# 카테고리칼럼 추가
sunstick_review_df['category'] = 'sunstic'

# 토큰화를 위해 결측치 채우기
sunstick_review_df.fillna("")

# 명사로 구분된는 단어중 인사이트에 의미없는 단어 리스트
sunstick_outword = ["때", '쌀', '요', '양', '형', '혼', '넘', '톤', '몰', '하', '연', '안', '폰', '티', '목', '립', '킴',
         '종', '전', '본', '가', '미', '다', '게', '아', '혀', '눈', '향', '함', '침', '맘', '후', '맥', '힘', '달', '겉', '선', '열', '땀',
         '날', '말', '밖', '여', '갈', '을', '못', '손', '물', '사', '폼', '입', '이', '당', '족', '파', '뒤', '볼', '코', '끔', '한', '톰',
         '업', '용', '콩', '떡', '허', '팁', '기', '뼈', '임', '끝', '감', '빛', '덧', '싱', '노', '칠', '품', '얼', '오', '위', '일', '틈',
         '리', '습', '속', '도', '뜸', '색', '어', '짜', '피', '발', '영', '남', '쌍', '야', '엉', '앞', '욥', '개', '귀', '썬', '골', '봄',
         '옆', '보', '팔', '인', '꿀', '그', '학', '괘', '은', '병', '해', '졸', '똥', '씀', '정', '두', '딸', '핑', '특', '알', '조', '흠',
         '팥', '합', '막', '잘', '난', '콘', '셀', '통', '누', '무', '더', '과', '팩', '화', '짐', '재', '지', '옷', '린', '모', '광', '글',
         '독', '애', '디', '션', '내', '매', '엽', '떼', '쌈', '뽕', '집', '능', '몸', '소', '추', '친', '완', '첨', '중', '곳', '쥬', '각',
         '느', '잇', '베', '챙', '겨', '서', '덤', '생', '마', '예', '세', '으', '염', '유', '나', '적', '러', '약', '새', '항', '상', '흰',
         '끼', '메', '컵', '악', '흑', '데', '음', '옴', '담', '획', '신', '로', '갑', '면', '단', '읍', '고', '곱', '띠', '엿', '넌', '냉',
         '백', '결', '존', '극', '총', '납', '라', '레', '래', '멜', '호', '턱', '타', '쿠', '뭐', '맛', '찬', '따', '녀', '살', '뺨', '뜰',
         '체', '절', '득', '드', '김', '규', '별', '변', '효', '운', '뭍', '길', '올', '죽', '순', '짝', '버', '릭', '겹', '끈', '냐', "요조",
         '편', '만', '밥', '식', '믈', '주', '풀', '갸', '부', '공', '혹', '력', '춤', '표', '배', '시', '밑', '낮', '진', '벗', '덕', '땅',
         '건', '릴', '돌', '착', '컷', '장', '비', '현', '불', '밤', '반', '니', '닷', '직', '처', '참', '벤', '박', '방', '국', '환', '쥐',
         '르', '룩', '암', '탓', '님', '홈', '패', '답', '차', '급', '점', '꿈', '에', '투', '숨', '강', '꽃', '랑', '까', '징', '걸', '센',
         '언', '왕', '엄', '찌', '링', '빠', '껌', '토', '카', '치', '질', '켄', '왜', '뜻', '구', '성', '좌', '우', '충', '거', '태', '증',
         '의', '저', '톱', '탁', '앎', '테', '섶', '림', '꼴', '팀', '액', '둑', '런', '탑', '룸', '싹', '락', '흉', '샘', '응', '송', '망',
         '홀', '줄', '낫', '필', '빙', '팝', '빅', '작', '문', '민', '실', '심', '텐', '죤', '울', '키', '값', '릉', '철', '동', '금', '팬',
         '홍', '창', '벡', '검', '깐', '칸', '휘', '율', '깃', '랩', '잎', '평', '숲', '논', '콜', '틱', '계', '류', '최', '축', '닐', '탄',
         '힐', '윤', '산', '므', '잠', '헌', '짓', '곰', '택', '낭', '덫', '몽', '역', '란', '들', '샹', '뇨', '핫', '범', '댐', '칩', '록',
         '외', '술', '빵', '궁', '뻘', '폴', '묘', '셤', '녜', '줌', '농', '샷', '삼', '찜', '와', '수', '먹', '펙', '초', '원', '폭', '혁',
         '믹', '법', '끌', '핵', '온', '닻', '청', '멱', '펄', '뤼', '량', '복', '잡', '묵', '칙', '솔', '맨', '잔', '컴', '자', '휴', '밭',
         '좀', '삽', '근', '접', '포', '썰', '땡', '탕', '붓', '펜', '탈', '벨', '솜', '댁', '램', '롤', '봉', '몬', '벽', '틀', '햄', '벼',
         '꾀', '멋', '씨', '껀', '젤', '즙', '활', '석', '꽈', '몫', '뱃', '옥', '녹', '촉', '킥', '엔', '볕', '멀', '쇼', '첩', '터', '겁',
         '젓', '굴', '행', '붕', '훔', '루', '놀', '론', '깨', '솥', '경', '출', '갭', '욕', '갱', '밴', '널', '털', '돗', '쾌', '죄', '붐',
         '선스틱', '썬스틱', '구매', '제품', '사용', '느낌', '생각', '모로', "개도", "데바르", "이면", "이건", "아악", "전이", "호가", "안보",
          "군가", "지나", "리한", "여물", "바쿠", "바르", "쓰기", "하주", "이랑"]

from tqdm.notebook import tqdm
import re
from konlpy.tag import Komoran


tqdm.pandas()

komoran = Komoran(userdic = "./data/user.dic")
def clean_text(text):# 이모티콘제거
    if isinstance(text, str):
        text = re.sub(r'[^\x00-\x7F\uAC00-\uD7AF]', '', text)
        text = re.sub(r'[!]', '', text)
        return text
    else:
        return ''  

def process_review(sentence):# 토큰화 진행(명사가져오기, outword_list의 단어 빼기)
    try:
        clean_sentence = clean_text(sentence) 
        tokens = komoran.pos(clean_sentence)
        nouns = [x[0] for x in tokens if x[1] in ("NNG", "NNP") and x[0] not in sunstic_outword]
        return nouns
    except UnicodeDecodeError as e:
        print(f"UnicodeDecodeError processing sentence: {sentence}")
        print(e)
        return []
    except Exception as e:
        print(f"Error processing sentence: {sentence}")
        print(e)
        return []


sunstick_review_df['processed_review'] = sunstick_review_df['review'].progress_apply(process_review)


# 파일 저장하기
sunstick_review_df.to_csv("./preprocessed_data/preprocessed_sunstick_data.csv",encoding="utf-8-sig",index=False)

## 4. Sunpatch

# 선패치 리뷰 데이터 불러오기
sunpatch_reviews_df = pd.read_csv("./crawling_data/sunpatch_reviews&buyrank.csv",encoding="utf-8")

# 선패치 링크 데이터 불러오기
sunpatch_link_df = pd.read_csv("./crawling_data/sunpatch_link.csv",encoding='utf-8')

# 브랜드명 리뷰데이터에 추가해주기
sunpatch_reviews_df = pd.merge(sunpatch_reviews_df, sunpatch_link_df[['product_name','brand_name']], on = 'product_name',how='left')

# 평점칼럼 전처리
sunpatch_reviews_df['star'] = sunpatch_reviews_df['star'].apply(lambda a: int(a.split(" ")[1][0]))

# 피부타입 칼럼 전처리
sunpatch_reviews_df['skin_type'] = sunpatch_reviews_df['skin_type'].apply(lambda x: eval(x)[0] if isinstance(eval(x), list) and len(eval(x)) > 0 else "")

# 카테고리열 추가
sunpatch_reviews_df['category']='sunpatch'

# 토큰화를 위해 결측치 채우기
sunpatch_reviews_df.fillna("")

# 명사로 구분된는 단어중 인사이트에 의미없는 단어 리스트
sunpatch_outword = ['이번', '제가', '해서', '느낌', '생각', '제품', '구매', '사용', '때', '패치', '전', '사', '합', '당', '갈', '왜', '한', '겨', 
           '편', '을', '김', '날', '안', '후','집', '넘', '칠', '면', '요', '위', '막', '이', '양', '낮', '라', '못', '선', '아', 
           '여', '어', '낙', '영', '첨', '일', '용', '함', '느', '걸', '극', '차', '부', '눈', '티', '흠', '산', '뒤', '고', '규', 
           '열', '력', '쌈', '보', '감', '질', '예', '글', '개', '뭐', '다', '애', '달', '황', '심', '만', '습', '오', '뜰','갑', 
           '피', '곳', '미', '와', '출', '샷', '시', '폭', '껀', '메', '암', '게', '니', '닷', '쌍', '입', '큐', '링', '디', '버', 
           '잘', '하', '유', '정', '난', '박', '길', '불', '드', '남', '띠', '랩', '답', '액', '서', '엄', '청', '주', '표', '가', 
           '끔', '배', '끝', '떼', '말', '패', '성', '종', '순', '분', '들', '법', '학', '악', '금', '진', '폴', '솔', '자', '잔', 
           '폰', '봉', '음', '밑', '상', '얼', '기', '색', '향', '존', '겉', '구', '꾀', '융', '의', '더', '삶', '광', '천', '귀', 
           '우', '파', '르', '직', '에', '최', '공', '씨', '땅', '품', '컷', '지', '득', '은', '리', '괘', '룸', '과', '추', '겁',
           '돌', '실', '알', '강', '녀', '도', '까', '덕', '틈', '두', '담', '조', '으', '봄', '랑', '접', '깨', '핑', '통', '엘', 
           '인', '류', '끈', '란', '비', '효', '흉', '혼', '풀', '형', '찬', '단', '홀', '엉', '꼴', '따', '임', '본', '매', '볕', 
           '밭', '포', '냉', '평', '테', '팁', '필', '계', '린', '쥐', '로', '총', '붐', '런', '쿠', '착', '림', '동', '뇨', '세', 
           '촉', '반', '모', '적', '각', '빰', '야', '급', '노', '놀', '탓', '업', '쥬', '운', '염', '제', '빠', '혀', '화', '낫', 
           '밴', '점', '토', '러', '석', '신', '톨', '완', '룰', '나', '절', '탈', '칼', '족', '소', '작', '중', '휴', '재', '래', 
           '졸', '삼', '능', '욥', '살', '힐', '방', '문', '처', '벗', '새', '내', '짐', '백', '뉴', '좀', '침', '첩', '마', '관', 
           '끼', '책', '국', '텐', '씀', '널', '맨', '톰', '궁', '션', '흑', '몰', '님', '레', '굴', '울', '줄', '저']

from tqdm.notebook import tqdm
import re
from konlpy.tag import Komoran


tqdm.pandas()

komoran = Komoran(userdic = "./data/user.dic")
def clean_text(text):# 이모티콘제거
    if isinstance(text, str):
        text = re.sub(r'[^\x00-\x7F\uAC00-\uD7AF]', '', text)
        text = re.sub(r'[!]', '', text)
        return text
    else:
        return ''  

def process_review(sentence):# 토큰화 진행(명사가져오기, outword_list의 단어 빼기)
    try:
        clean_sentence = clean_text(sentence)  
        tokens = komoran.pos(clean_sentence)
        nouns = [x[0] for x in tokens if x[1] in ("NNG", "NNP") and x[0] not in sunpatch_outword]
        return nouns
    except UnicodeDecodeError as e:
        print(f"UnicodeDecodeError processing sentence: {sentence}")
        print(e)
        return []
    except Exception as e:
        print(f"Error processing sentence: {sentence}")
        print(e)
        return []


sunpatch_review_df['processed_review'] = sunpatch_review_df['review'].progress_apply(process_review)


#파일저장
sunpatch_review_df.to_csv("./preprocessed_data/preprocessed_sunpatch_data.csv",encoding="utf-8-sig",index=False)

## 5. Sunspray

# 선스프레이 리뷰 데이터 불러오기
sunspray_reviews_df = pd.read_csv("./crawling_data/sunpray_reviews&buyrank.csv",encoding="utf-8")

# 선스프레이 링크 데이터 불러오기
sunspray_link_df = pd.read_csv("./crawling_data/sunpray_link.csv",encoding='utf-8')

# 브랜드명 리뷰데이터에 추가해주기
sunspray_reviews_df = pd.merge(sunspray_reviews_df, sunspray_link_df[['product_name','brand_name']], on = 'product_name',how='left')

# 평점칼럼 전처리
sunspray_reviews_df['star'] = sunspray_reviews_df['star'].apply(lambda a: int(a.split(" ")[1][0]))

# 피부타입 칼럼 전처리
sunspray_reviews_df['skin_type'] = sunspray_reviews_df['skin_type'].apply(lambda x: eval(x)[0] if isinstance(eval(x), list) and len(eval(x)) > 0 else "")

#카테고리열 추가
sunspray_reviews_df['category']='sunspray'

# 토큰화를 위해 결측치 채우기
sunspray_reviews_df.fillna("")

# 명사로 구분된는 단어중 인사이트에 의미없는 단어 리스트
sunspray_outword = ['이번', '제가', '해서', '느낌', '생각', '제품', '구매', '사용', '때', '전', '사', '합', '당', '갈', '왜', '한', '겨', 
           '편', '을', '김', '날', '안', '후','집', '넘', '칠', '면', '요', '위', '막', '이', '양', '낮', '라', '못', '선', '아', 
           '여', '어', '낙', '영', '첨', '일', '용', '함', '느', '걸', '극', '차', '부', '눈', '티', '흠', '산', '뒤', '고', '규', 
           '열', '력', '쌈', '보', '감', '질', '예', '글', '개', '뭐', '다', '애', '달', '황', '심', '만', '습', '오', '뜰','갑', 
           '피', '곳', '미', '와', '출', '샷', '시', '폭', '껀', '메', '암', '게', '니', '닷', '쌍', '입', '큐', '링', '디', '버', 
           '잘', '하', '유', '정', '난', '박', '길', '불', '드', '남', '띠', '랩', '답', '액', '서', '엄', '청', '주', '표', '가', 
           '끔', '배', '끝', '떼', '말', '패', '성', '종', '순', '분', '들', '법', '학', '악', '금', '진', '폴', '솔', '자', '잔', 
           '폰', '봉', '음', '밑', '상', '얼', '기', '색', '향', '존', '겉', '구', '꾀', '융', '의', '더', '삶', '광', '천', '귀', 
           '우', '파', '르', '직', '에', '최', '공', '씨', '땅', '품', '컷', '지', '득', '은', '리', '괘', '룸', '과', '추', '겁',
           '돌', '실', '알', '강', '녀', '도', '까', '덕', '틈', '두', '담', '조', '으', '봄', '랑', '접', '깨', '핑', '통', '엘', 
           '인', '류', '끈', '란', '비', '효', '흉', '혼', '풀', '형', '찬', '단', '홀', '엉', '꼴', '따', '임', '본', '매', '볕', 
           '밭', '포', '냉', '평', '테', '팁', '필', '계', '린', '쥐', '로', '총', '붐', '런', '쿠', '착', '림', '동', '뇨', '세', 
           '촉', '반', '모', '적', '각', '빰', '야', '급', '노', '놀', '탓', '업', '쥬', '운', '염', '제', '빠', '혀', '화', '낫', 
           '밴', '점', '토', '러', '석', '신', '톨', '완', '룰', '나', '절', '탈', '칼', '족', '소', '작', '중', '휴', '재', '래', 
           '졸', '삼', '능', '욥', '살', '힐', '방', '문', '처', '벗', '새', '내', '짐', '백', '뉴', '좀', '침', '첩', '마', '관', 
           '끼', '책', '국', '텐', '씀', '널', '맨', '톰', '궁', '션', '흑', '몰', '님', '레', '굴', '울', '줄', '저']

from tqdm.notebook import tqdm
import re
from konlpy.tag import Komoran


tqdm.pandas()

komoran = Komoran(userdic = "./data/user.dic")
def clean_text(text):# 이모티콘제거
    if isinstance(text, str):
        text = re.sub(r'[^\x00-\x7F\uAC00-\uD7AF]', '', text)
        text = re.sub(r'[!]', '', text)
        return text
    else:
        return ''  

def process_review(sentence):# 토큰화 진행(명사가져오기, outword_list의 단어 빼기)
    try:
        clean_sentence = clean_text(sentence) 
        tokens = komoran.pos(clean_sentence)
        nouns = [x[0] for x in tokens if x[1] in ("NNG", "NNP") and x[0] not in sunspray_outword]
        return nouns
    except UnicodeDecodeError as e:
        print(f"UnicodeDecodeError processing sentence: {sentence}")
        print(e)
        return []
    except Exception as e:
        print(f"Error processing sentence: {sentence}")
        print(e)
        return []


sunspray_reviews_df['processed_review'] = sunspray_reviews_df['review'].progress_apply(process_review)


#파일저장
sunspray_review_df.to_csv("./preprocessed_data/preprocessed_sunspray_data.csv",encoding="utf-8-sig",index=False)
