==> 4-1. category_Top&Bottom_10_analysis.ipynb <==
# 4-1. 카테고리별 리뷰 평균 평점 Top & Bottom 10 목록 뽑기

1. **선크림, 선스틱, 선쿠션** 카테고리 
- 제품 수와 제품별 리뷰 수가 충분히 많음
- 데이터 신뢰성 확보를 위해 리뷰 수 100개 이상인 제품 안에서 상하위 제품 각 10개씩 선정
2. **선패치, 선스프레이** 카테고리
- 전체 제품 수가 20개 미만인 관계로 전체 제품에서 상위, 하위 제품 각 3개씩을 선정

import pandas as pd
import matplotlib.pyplot as plt
from matplotlib import font_manager, rc

plt.rcParams['font.family'] = 'NanumGothic'

## 선크림 평균 평점 상하위 제품
우선 충분한 데이터 표본이 신뢰성이 높다고 판단되어 리뷰댓글수가 100개 이상인 제품만 추려 확인하였습니다.

suncream_review_df = pd.read_csv("./preprocessed_data/preprocessed_suncream_data.csv",encoding="utf-8",low_memory=False)

product_review_counts = suncream_review_df.groupby('product_name')['review'].count()

filtered_products = product_review_counts[product_review_counts >= 100].index
star_mean_by_filtered_product = suncream_review_df[suncream_review_df['product_name'].isin(filtered_products)].groupby('product_name')['star'].mean()

top10_products = star_mean_by_filtered_product.nlargest(10)

top10_products.index

bottom10_products = star_mean_by_filtered_product.nsmallest(10)

bottom10_products.sort_values(ascending=True).index

## 선스틱 평균 평점 상하위 제품

sunstic_review_df = pd.read_csv("./preprocessed_data/preprocessed_sunstic_data.csv",encoding="utf-8",low_memory=False)

product_review_counts = sunstic_review_df.groupby('product_name')['review'].count()

filtered_products = product_review_counts[product_review_counts >= 100].index
star_mean_by_filtered_product = sunstic_review_df[sunstic_review_df['product_name'].isin(filtered_products)].groupby('product_name')['star'].mean()

top10_products = star_mean_by_filtered_product.nlargest(10)

top10_products.index

bottom10_products = star_mean_by_filtered_product.nsmallest(10)

bottom10_products.sort_values(ascending=True).index

## 선쿠션 평균 평점 상하위 제품

suncushion_review_df = pd.read_csv("./preprocessed_data/preprocessed_suncushion_data.csv",encoding="utf-8")

product_review_counts = suncushion_review_df.groupby('product_name')['review'].count()

filtered_products = product_review_counts[product_review_counts >= 100].index
star_mean_by_filtered_product = suncushion_review_df[suncushion_review_df['product_name'].isin(filtered_products)].groupby('product_name')['star'].mean()

top10_products = star_mean_by_filtered_product.nlargest(10)

top10_products.index

bottom10_products = star_mean_by_filtered_product.nsmallest(10)

bottom10_products.sort_values(ascending=True).index

## 선패치 평균 평점 상하위 제품

sunpatch_review_df = pd.read_csv("./preprocessed_data/preprocessed_sunpatch_data.csv",encoding="utf-8")

product_review_counts = sunpatch_review_df.groupby('product_name')['review'].count()

filtered_products = product_review_counts[product_review_counts >= 100].index
star_mean_by_filtered_product = sunpatch_review_df[sunpatch_review_df['product_name'].isin(filtered_products)].groupby('product_name')['star'].mean()

top3_products = star_mean_by_filtered_product.nlargest(3)

top3_products.index

bottom3_products = star_mean_by_filtered_product.nsmallest(3)

bottom3_products.index

## 선스프레이 평균 평점 상하위 제품

sunspray_review_df = pd.read_csv("./preprocessed_data/preprocessed_sunspray_data.csv",encoding="utf-8")

product_review_counts = sunspray_review_df.groupby('product_name')['review'].count()

filtered_products = product_review_counts[product_review_counts >= 100].index
star_mean_by_filtered_product = sunspray_review_df[sunspray_review_df['product_name'].isin(filtered_products)].groupby('product_name')['star'].mean()

top3_products = star_mean_by_filtered_product.nlargest(3)

top3_products.index

bottom3_products = star_mean_by_filtered_product.nsmallest(3)

bottom3_products.index
