==> 6. review_embedding_and_pinecone_upload.ipynb <==
# 6. 리뷰 임베딩 벡터 변환 및 파인콘 업로드

from openai import OpenAI
from getpass import getpass

api_key=getpass("OPENAI_API_KEY")

from tqdm import tqdm

client = OpenAI(api_key=api_key)

import pandas as pd

suncream_df = pd.read_csv('./data/sunpatch_final_all.csv', encoding='utf-8')

sunspray_df = pd.read_csv('./data/sunspray_final_all.csv', encoding='utf-8')

sunspray_df

sunspray_df = sunspray_df.drop(['title', 'brand_name', 'category', 'star', 'star_numeric', 'sales_ranking', 'processed_review'], axis = 1)

tqdm.pandas()
sunspray_df["skin_type"] = sunspray_df["skin_type"].progress_apply(lambda x: eval(x))

tmp = []
for i, type_ in enumerate(sunspray_df["skin_type"]):
    tmp.append(" ".join(type_))

sunpatch_df = sunpatch_df.drop(["skin_type"], axis = 1)

sunspray_df["skintype"] = tmp

sunspray_df

sunspray_df["review"] = sunspray_df["review"] + " " + sunspray_df["skintype"]

sunspray_df = sunspray_df.drop(["skintype"], axis = 1)

sample = df.iloc[3]
sample_review = sample.review


sample_review

def extract_embedding(text_list):
    response=client.embeddings.create(
        input=text_list,
        model='text-embedding-3-large',
    )
    embedding_list = [x.embedding for x in response.data]
    return embedding_list

from tqdm.notebook import tqdm

embeddings=[]
batchsize=10
for i in tqdm(range(0, len(sunspray_df), 10)):
    batch=sunspray_df.iloc[i:i+batchsize]
    text_list = batch['review'].tolist()
    batch_embedding=extract_embedding(text_list)
    embeddings.extend(batch_embedding)

sunspray_df["embedding"] = embeddings

len(embeddings)

len(embeddings[0])

sunspray_df

sunpatch_df.iloc[0]["review"]

tmp = []

for type_ in sunspray_df["skin_type"]:
    if not type_:
        tmp.append("empty")
    else:
        tmp.append(type_)

sunspray_df = sunspray_df.drop(["skin_type"], axis = 1)

sunspray_df["skin_type"] = tmp

sunspray_df

# PINECONE

from pinecone import Pinecone

pinecone_api_key = getpass("PINECONE_API_KEY")

pc = Pinecone(api_key = pinecone_api_key)
index = pc.Index("review")

index.describe_index_stats()

sunpatch_df = sunpatch_df[["page", "product_name", "review", "skin_type", "embedding"]]

sunspray_df

sunpatch_df['review']

output_path = './data/sunpatch_vector.csv'
sunpatch_df.to_csv(output_path, index=False, encoding = 'utf-8-sig')

batchsize = 100

for i in tqdm(range(0, len(sunspray_df), batchsize)):
    batch = sunspray_df.iloc[i:i+batchsize]
    vectors = []
    for j in range(len(batch)):
        row = batch.iloc[j]
        id_ = str(row.name)
        embedding = row["embedding"]
        metadata = row.drop(["embedding"]).to_dict()
        vectors.append((id_, embedding, metadata))
    response = index.upsert(vectors)

type_ = input("피부타입 :")

query = input("입력 :")

full_query = f"""
유저가 사용하고 싶은 제품을 설명합니다.
이를 참고하여 유저에게 제품을 추천해주세요.
```
피부타입 : {type_}
사용하고싶은 제품 : {query}
```
""".strip()

query_embedding = extract_embedding([full_query])[0]

result = index.query(
    vector = query_embedding,
    top_k = 3,
    include_metadata = True,
    filter={
    "skin_type" : type_ or "empty"
    }
)

result

# AI 추천

def generate_prompt(query, items):
    prompt = f"""
    유저가 사용하고 싶은 제품에 대한 설명와 이에 대한 추천 결과가 주어집니다.
    유저의 입력과 각 추천 결과 제품이름, 피부타입, 리뷰 등을 참고하여 추천사를 작성하세요.
    당신에 대한 소개를 먼저 하고, 친절한 말투로 작성해주세요.
    중간 중간 이모지를 적절히 사용해주세요.

    ```
    {items}
    ```
    """
    return prompt

matches = result["matches"]

metadata_list = [x["metadata"] for x in matches]

item_list = [{
    "제품이름" : x["product_name"],
    "피부타입" : x["skin_type"],
    "리뷰" : x["review"]
} for x in metadata_list]

import json

prompt = generate_prompt(
    query = query,
    items = json.dumps(item_list, indent = 2, ensure_ascii = False)
)

print(prompt)

def request_chat_completion(prompt):
    response = client.chat.completions.create(
        model = "gpt-4o",
        messages = [
            {"role" : "system", "content" : "당신은 제품을 추천해주는 올리브영 점장입니다."},
            {"role" : "user", "content" : prompt}
        ],
        stream = True
    )
    return response

def print_streaming_response(response):
    for chunk in response:
        delta = chunk.choices[0].delta
        if delta.content:
            print(delta.content, end = "")

response = request_chat_completion(prompt)
print_streaming_response(response)

# 정리

1. 리뷰, 스킨타입, 고민거리 각각 임베딩 시킨다.  
 - 리뷰 : 영어로 변환 → 이모지는 구글에서 표시는 가능해서 안지워도 됨, 리뷰 없는 데이터는 삭제  
 - 고민거리 : 영어로 변환  
 - 스킨타입 : 영어로 변환, 첫번째 요소만 남김  
    → 스킨타입 칼럼은 문자열 타입이라 리스트 변환 필요  
    → 스킨타입에서 고민거리 분리    
2. 제품별로 그룹화해서 임베딩 값 평균을 계산한다.

3. 제품별 리뷰, 스킨타입을 결합시킨다.

4. 사용자에게 스킨타입, 추가요청사항을 받는다.
 - 입력받은 값 영어로 변환  

5. 사용자에게 입력받은 값을 각각 임베딩 시켜서 결합시킨다.

6. 코사인유사도를 사용하여 두 벡터간 유사성을 측정한다.

7. 유사도가 높은 상위 N개 제품을 사용자에게 추천한다.
